[EI-128] [web] rename: Rename dialog to alert-dialog component
